removed activity container
